Redistrubute functions among the vi1 rendering files
